add role_checking

+++ pkg/controllers/lesson_material.py
@@ -2,14 +2,12 @@ import json
 
 from fastapi import APIRouter, status, Depends, HTTPException, File, UploadFile
 from starlette.responses import Response, JSONResponse
+from pathlib import Path
+import shutil
 
 from pkg.controllers.middlewares import get_current_user
 from logger.logger import logger
 from utils.auth import TokenPayload
-
-from pathlib import Path
-import shutil
-
 from pkg.services import lesson_material as material_service
 
 
@@ -28,8 +26,12 @@ router = APIRouter()
 
 
 # CREATE
-@router.post("/lesson-materials/{lesson_id}", summary="Upload a file", tags=["lesson-materials"])
-def upload_file(lesson_id: int, file: UploadFile = File(...)):
+@router.post("/lesson-materials/{lesson_id}", summary="Upload a file", tags=["lesson_materials"])
+def upload_file(lesson_id: int, file: UploadFile = File(...), payload: TokenPayload=Depends(get_current_user)):
+    role_id = payload.role_id
+    if role_id == 1:
+        return Response(json.dumps({"error": "only admins and mentors can add lesson materials"}),
+                        status_code=status.HTTP_403_FORBIDDEN)
     
     if file.size > MAX_ALLOWED_SIZE:
         return {"error": "File size should be less or equal to 5mb"}
@@ -45,7 +47,7 @@ def upload_file(lesson_id: int, file: UploadFile = File(...)):
 
 # READ
 @router.get("/lesson-materials/{lesson_id}", summary="Get list of all lesson_materials", tags=["lesson_materials"])
-def get_all_materials(lesson_id: int):  #payload: TokenPayload=Depends(get_current_user)):
+def get_all_materials(lesson_id: int, payload: TokenPayload=Depends(get_current_user)):
     try:
         materials = material_service.get_all_materials(lesson_id)
         return materials
@@ -54,16 +56,22 @@ def get_all_materials(lesson_id: int):  #payload: TokenPayload=Depends(get_curre
 
 
 @router.get("/lesson-materials/file/{file_id}", summary="Get lesson material by id", tags=["lesson_materials"])
-def get_material_by_id(file_id: int):
+def get_material_by_id(file_id: int, payload: TokenPayload=Depends(get_current_user)):
     try:
         lesson_material = material_service.get_material_by_id(file_id) 
         return JSONResponse({"lesson_material path": lesson_material}, status_code=status.HTTP_200_OK)
     except Exception as e:
         return {"message": e.args}
     
+
 # UPDATE
 @router.put("/lesson-materials/file/{file_id}", summary="Update lesson material by id", tags=["lesson_materials"])
-def update_file(file_id: int, file: UploadFile = File(...)):
+def update_file(file_id: int, file: UploadFile = File(...), payload: TokenPayload = Depends(get_current_user)):
+    role_id = payload.role_id
+    if role_id == 1:
+        return Response(json.dumps({"error": "only admins and mentors can update lesson materials"}),
+                        status_code=status.HTTP_403_FORBIDDEN)
+    
     if file.size > MAX_ALLOWED_SIZE:
         return {"error": "File size should be less or equal to 5mb"}
     
@@ -79,7 +87,12 @@ def update_file(file_id: int, file: UploadFile = File(...)):
 
 # DELETE
 @router.delete("/lesson-materials/file/{file_id}", summary="Delete lesson material by id", tags=["lesson_materials"])
-def delete_file(file_id):
+def delete_file(file_id, payload: TokenPayload = Depends(get_current_user)):
+    role_id = payload.role_id
+    if role_id == 1:
+        return Response(json.dumps({"error": "only admins and mentors can update lesson materials"}),
+                        status_code=status.HTTP_403_FORBIDDEN)
+    
     if material_service.delete_file(file_id):
         return JSONResponse({
             "message": "successfully deleted the file"
